requirements.txt
Flask==2.3.2
google-generativeai==0.5.2
python-pptx==0.6.21
fpdf==1.7.2
python-dotenv==1.0.0

# Dev tools (optional)
flake8==6.0.0
pytest==7.4.0
